+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/conf/WaggleDanceConfiguration.java
@@ -45,6 +45,7 @@ public class WaggleDanceConfiguration {
   private int thriftServerStopTimeoutValInSeconds = 60;
   private int thriftServerRequestTimeout = 20;
   private TimeUnit thriftServerRequestTimeoutUnit = TimeUnit.SECONDS;
+  private int clientPoolMaxTotalPerKey = 100;
 
   public Integer getPort() {
     return port;
@@ -119,4 +120,11 @@ public class WaggleDanceConfiguration {
     this.thriftServerRequestTimeoutUnit = thriftServerRequestTimeoutUnit;
   }
 
+  public int getClientPoolMaxTotalPerKey() {
+    return clientPoolMaxTotalPerKey;
+  }
+
+  public void setClientPoolMaxTotalPerKey(int clientPoolMaxTotalPerKey) {
+    this.clientPoolMaxTotalPerKey = clientPoolMaxTotalPerKey;
+  }
 }