+++ waggle-dance-rest/src/test/java/com/hotels/bdp/waggledance/rest/endpoint/FederationsAdminControllerTest.java
@@ -15,7 +15,7 @@
  */
 package com.hotels.bdp.waggledance.rest.endpoint;
 
-import static org.mockito.Matchers.anyString;
+import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
@@ -69,7 +69,7 @@ public class FederationsAdminControllerTest {
   public void setUp() {
     MockitoAnnotations.initMocks(this);
     mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).build();
-    when(federationStatusService.checkStatus(anyString())).thenReturn(MetaStoreStatus.AVAILABLE);
+    when(federationStatusService.checkStatus(any(AbstractMetaStore.class))).thenReturn(MetaStoreStatus.AVAILABLE);
     metastore.setStatus(MetaStoreStatus.AVAILABLE);
   }