+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## TBD
+### Fixed
+* Metastore status didn't work for tunneled connections. #issue-34
+
+
 ## [2.2.1] 2017-10-30
 ### Fixed
 * Metastore status was missing from api/admin/federations rest endpoint. #issue-29